Add test_mode

+++ js/components/subheader.js
@@ -1,12 +1,16 @@
 window.onload = function () {
-  const token = sessionStorage.getItem("token");
-  if (!token) {
-    alert(`No token found! Redirecting to login page.`);
-    sessionStorage.clear();
-    window.location.href = "https://apmaries.github.io/wpt/index.html";
+  if (!sessionStorage.getItem("test_mode")) {
+    const token = sessionStorage.getItem("token");
+    if (!token) {
+      alert(`No token found! Redirecting to login page.`);
+      sessionStorage.clear();
+      window.location.href = "https://apmaries.github.io/wpt/index.html";
+    }
+    const authText = document.getElementById("authenticatedSubHeader");
+    const orgName = sessionStorage.getItem("org_name");
+    const userName = sessionStorage.getItem("user_name");
+    authText.innerHTML = `${userName} authenticated in: ${orgName}`;
+  } else {
+    console.log("WPT: Test mode active");
   }
-  const authText = document.getElementById("authenticatedSubHeader");
-  const orgName = sessionStorage.getItem("org_name");
-  const userName = sessionStorage.getItem("user_name");
-  authText.innerHTML = `${userName} authenticated in: ${orgName}`;
 };
